+++ client/src/components/game/entities.js
@@ -1,0 +1,195 @@
+import { ENEMY_SIZE, ENEMY_HITBOX_SIZE, ENEMY_SPEED, MONEY_SIZE, PLAYER_SIZE, MAX_HEALTH, PLAYER_SPEED, GUNS } from './constants';
+
+export class Enemy {
+    constructor(x, y, ctx, canvas) {
+        this.x = x;
+        this.y = y;
+        this.health = 50;
+        this.ctx = ctx;
+        this.canvas = canvas;
+        this.size = ENEMY_SIZE;
+        this.hitboxSize = ENEMY_HITBOX_SIZE;
+    }
+
+    draw() {
+        this.ctx.fillStyle = '#ff4444';
+        this.ctx.beginPath();
+        this.ctx.arc(this.x, this.y, this.size/2, 0, Math.PI * 2);
+        this.ctx.fill();
+
+        // Health bar
+        this.ctx.fillStyle = 'white';
+        this.ctx.fillRect(this.x - 20, this.y - 30, 40, 4);
+        this.ctx.fillStyle = 'red';
+        this.ctx.fillRect(this.x - 20, this.y - 30, (this.health / 50) * 40, 4);
+    }
+
+    moveTowards(targetX, targetY) {
+        const angle = Math.atan2(targetY - this.y, targetX - this.x);
+        this.x += Math.cos(angle) * ENEMY_SPEED;
+        this.y += Math.sin(angle) * ENEMY_SPEED;
+    }
+}
+
+export class MoneyDrop {
+    constructor(x, y, amount, ctx) {
+        this.x = x;
+        this.y = y;
+        this.amount = amount;
+        this.ctx = ctx;
+    }
+
+    draw() {
+        this.ctx.fillStyle = '#FFD700';
+        this.ctx.beginPath();
+        this.ctx.arc(this.x, this.y, MONEY_SIZE/2, 0, Math.PI * 2);
+        this.ctx.fill();
+        this.ctx.fillStyle = 'black';
+        this.ctx.font = '12px Arial';
+        this.ctx.textAlign = 'center';
+        this.ctx.fillText('$' + this.amount, this.x, this.y + 4);
+    }
+}
+
+export class Player {
+    constructor(x, y, color, controls, ctx, canvas) {
+        this.x = x;
+        this.y = y;
+        this.color = color;
+        this.controls = controls;
+        this.health = MAX_HEALTH;
+        this.money = 100000;
+        this.bullets = [];
+        this.lastShot = 0;
+        this.ctx = ctx;
+        this.canvas = canvas;
+        this.direction = 'down';
+        this.lastDamageTime = 0;
+        this.lastHitTime = 0;
+        this.isFlashing = false;
+    }
+
+    getCurrentGun() {
+        return GUNS[window.currentGun];
+    }
+
+    draw() {
+        // Draw player body with damage flash effect
+        this.ctx.fillStyle = this.isFlashing ? '#ff0000' : this.color;
+        this.ctx.beginPath();
+        this.ctx.arc(this.x, this.y, PLAYER_SIZE/2, 0, Math.PI * 2);
+        this.ctx.fill();
+
+        // Draw direction indicator
+        this.ctx.fillStyle = '#fff';
+        let indicatorX = this.x;
+        let indicatorY = this.y;
+        const offset = PLAYER_SIZE/2 - 5;
+        
+        switch(this.direction) {
+            case 'right': indicatorX += offset; break;
+            case 'left': indicatorX -= offset; break;
+            case 'up': indicatorY -= offset; break;
+            case 'down': indicatorY += offset; break;
+        }
+        
+        this.ctx.beginPath();
+        this.ctx.arc(indicatorX, indicatorY, 5, 0, Math.PI * 2);
+        this.ctx.fill();
+
+        // Draw bullets
+        this.bullets.forEach(bullet => {
+            this.ctx.fillStyle = this.getCurrentGun().color;
+            this.ctx.beginPath();
+            this.ctx.arc(bullet.x, bullet.y, this.getCurrentGun().bulletSize, 0, Math.PI * 2);
+            this.ctx.fill();
+        });
+
+        // Draw current gun name
+        this.ctx.fillStyle = this.color;
+        this.ctx.font = '12px Arial';
+        this.ctx.textAlign = 'center';
+        this.ctx.fillText(this.getCurrentGun().name, this.x, this.y + 40);
+    }
+
+    move() {
+        if (this.controls.up && this.y > PLAYER_SIZE/2) {
+            this.y -= PLAYER_SPEED;
+            this.direction = 'up';
+        }
+        if (this.controls.down && this.y < this.canvas.height - PLAYER_SIZE/2) {
+            this.y += PLAYER_SPEED;
+            this.direction = 'down';
+        }
+        if (this.controls.left && this.x > PLAYER_SIZE/2) {
+            this.x -= PLAYER_SPEED;
+            this.direction = 'left';
+        }
+        if (this.controls.right && this.x < this.canvas.width - PLAYER_SIZE/2) {
+            this.x += PLAYER_SPEED;
+            this.direction = 'right';
+        }
+    }
+
+    shoot() {
+        const currentTime = Date.now();
+        const gun = this.getCurrentGun();
+        if (this.controls.shoot && currentTime - this.lastShot > gun.cooldown) {
+            let dx = 0, dy = 0;
+            switch(this.direction) {
+                case 'right': dx = 1; dy = 0; break;
+                case 'left': dx = -1; dy = 0; break;
+                case 'up': dx = 0; dy = -1; break;
+                case 'down': dx = 0; dy = 1; break;
+            }
+            
+            if (gun.name === 'Shotgun') {
+                for (let i = -1; i <= 1; i++) {
+                    const spread = i * 0.2;
+                    this.bullets.push({
+                        x: this.x,
+                        y: this.y,
+                        dx: dx + (dy * spread),
+                        dy: dy + (dx * spread)
+                    });
+                }
+            } else {
+                this.bullets.push({
+                    x: this.x,
+                    y: this.y,
+                    dx: dx,
+                    dy: dy
+                });
+            }
+            
+            this.lastShot = currentTime;
+        }
+    }
+
+    updateBullets(enemies) {
+        for (let i = this.bullets.length - 1; i >= 0; i--) {
+            const bullet = this.bullets[i];
+            bullet.x += bullet.dx * this.getCurrentGun().bulletSpeed;
+            bullet.y += bullet.dy * this.getCurrentGun().bulletSpeed;
+            
+            if (bullet.x < 0 || bullet.x > this.canvas.width || 
+                bullet.y < 0 || bullet.y > this.canvas.height) {
+                this.bullets.splice(i, 1);
+                continue;
+            }
+
+            for (let j = enemies.length - 1; j >= 0; j--) {
+                const enemy = enemies[j];
+                const dx = bullet.x - enemy.x;
+                const dy = bullet.y - enemy.y;
+                const distance = Math.sqrt(dx * dx + dy * dy);
+
+                if (distance < (enemy.hitboxSize/2 + this.getCurrentGun().bulletSize)) {
+                    enemy.health -= this.getCurrentGun().damage;
+                    this.bullets.splice(i, 1);
+                    break;
+                }
+            }
+        }
+    }
+}
